HomeWork3
